# ground-control/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Ground Control management API. Bearer authentication is the recommended workflow for interactive clients. Basic authentication remains supported on authenticated human endpoints for automation compatibility.",
    "title": "Harbor Satellite Ground Control API",
    "version": "1.0"
  },
  "basePath": "/",
  "paths": {
    "/api/groups/sync": {
      "post": {
        "description": "Creates the group described in the state artifact, updates the projects of all robot accounts attached to its satellites, ensures the \"satellite\" project exists in Harbor, and writes the state artifact. Requires an authenticated user. Bearer auth is recommended; Basic auth remains supported for automation.",
        "tags": [
          "groups"
        ],
        "summary": "Create or update a group from a state artifact",
        "operationId": "syncGroup",
        "parameters": [
          {
            "name": "state",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StateArtifact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group created and state synced",
            "schema": {
              "$ref": "#/definitions/Group"
            }
          },
          "400": {
            "description": "Invalid request body",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "502": {
            "description": "Upstream Harbor error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "description": "Bearer auth is recommended. Basic auth remains supported.",
        "tags": [
          "users"
        ],
        "summary": "List users",
        "operationId": "listUsers",
        "responses": {
          "200": {
            "description": "Users listed",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      },
      "post": {
        "description": "Requires a system admin user. Bearer auth is recommended.",
        "tags": [
          "users"
        ],
        "summary": "Create a user",
        "operationId": "createUser",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "User already exists",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/api/users/password": {
      "patch": {
        "description": "Bearer auth is recommended. Basic auth remains supported.",
        "tags": [
          "users"
        ],
        "summary": "Change the current user's password",
        "operationId": "changeOwnPassword",
        "parameters": [
          {
            "name": "password",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangeOwnPasswordRequest"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Password changed and sessions invalidated"
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized or current password incorrect",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/api/users/{username}": {
      "get": {
        "description": "Bearer auth is recommended. Basic auth remains supported.",
        "tags": [
          "users"
        ],
        "summary": "Get a user by username",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "User found",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      },
      "delete": {
        "description": "Requires a system admin user. Bearer auth is recommended.",
        "tags": [
          "users"
        ],
        "summary": "Delete a user",
        "operationId": "deleteUser",
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      },
      "parameters": [
        {
          "type": "string",
          "name": "username",
          "in": "path",
          "required": true
        }
      ]
    },
    "/api/users/{username}/password": {
      "patch": {
        "description": "Requires a system admin user. Bearer auth is recommended.",
        "tags": [
          "users"
        ],
        "summary": "Change another user's password",
        "operationId": "changeUserPassword",
        "parameters": [
          {
            "name": "password",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangeUserPasswordRequest"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Password changed and sessions invalidated"
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      },
      "parameters": [
        {
          "type": "string",
          "name": "username",
          "in": "path",
          "required": true
        }
      ]
    },
...
    "/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create a session token",
        "operationId": "login",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session created",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          },
          "400": {
            "description": "Invalid request body",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Artifact": {
      "type": "object",
      "properties": {
        "deleted": {
          "type": "boolean"
        },
        "digest": {
          "type": "string"
        },
        "labels": {
          "description": "Free-form label metadata; passed through unchanged.",
          "type": "object"
        },
        "repository": {
          "type": "string"
        },
        "tag": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "type": {
          "type": "string"
        }
      }
    },
    "ChangeOwnPasswordRequest": {
      "type": "object",
      "required": [
        "current_password",
        "new_password"
      ],
      "properties": {
        "current_password": {
          "type": "string",
          "format": "password"
        },
        "new_password": {
          "type": "string",
          "format": "password"
        }
      }
    },
    "ChangeUserPasswordRequest": {
      "type": "object",
      "required": [
        "new_password"
      ],
      "properties": {
        "new_password": {
          "type": "string",
          "format": "password"
        }
      }
    },
    "CreateUserRequest": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "password": {
          "type": "string",
          "format": "password"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "required": [
        "error"
      ],
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "Group": {
      "type": "object",
      "required": [
        "id",
        "group_name",
        "registry_url",
        "projects",
        "created_at",
        "updated_at"
      ],
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "group_name": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "projects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "registry_url": {
          "type": "string"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "LoginRequest": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "password": {
          "type": "string",
          "format": "password"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "LoginResponse": {
      "type": "object",
      "required": [
        "token",
        "expires_at"
      ],
      "properties": {
        "expires_at": {
          "type": "string",
          "format": "date-time"
        },
        "token": {
          "type": "string"
        }
      }
    },
...
    "StateArtifact": {
      "type": "object",
      "properties": {
        "artifacts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Artifact"
          }
        },
        "group": {
          "type": "string"
        },
        "registry": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "required": [
        "id",
        "username",
        "role",
        "created_at"
      ],
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "role": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "Users": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/User"
      }
    }
  },
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    },
    "BearerAuth": {
      "description": "Use `Bearer \u003ctoken\u003e`.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}