# ground-control/gcctl/apiclient/generated/models/artifact.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// Artifact artifact
//
// swagger:model Artifact
type Artifact struct {

	// deleted
	Deleted bool `json:"deleted,omitempty"`

	// digest
	Digest string `json:"digest,omitempty"`

	// Free-form label metadata; passed through unchanged.
	Labels any `json:"labels,omitempty"`

	// repository
	Repository string `json:"repository,omitempty"`

	// tag
	Tag []string `json:"tag"`

	// type
	Type string `json:"type,omitempty"`
}

// Validate validates this artifact
func (m *Artifact) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this artifact based on context it is used
func (m *Artifact) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Artifact) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Artifact) UnmarshalBinary(b []byte) error {
	var res Artifact
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}